requirements.txt
Adafruit-Blinka==8.38.0
Adafruit-CharLCD==1.1.1
adafruit-circuitpython-74hc595==1.4.4
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-charlcd==3.4.10
adafruit-circuitpython-connectionmanager==1.0.1
adafruit-circuitpython-mcp230xx==2.5.13
adafruit-circuitpython-requests==3.2.3
adafruit-circuitpython-typing==1.10.3
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.62.0
Adafruit-PureIO==1.1.11
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
assure==0.0.6
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorzero==2.0
drivers==0.0.5
filelock==3.13.3
fleep==1.0.1
frozenlist==1.4.1
fsspec==2024.3.1
gpiozero==2.0.1
huggingface_hub==0.22.2
idna==3.6
iniconfig==2.0.0
mfrc522==0.0.7
mmry==0.0.8
multidict==6.0.5
mysql-connector-python==8.3.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pluggy==1.4.0
pyftdi==0.55.0
PyJWT==2.8.0
pypdf==4.1.0
pyserial==3.5
pytest==8.1.1
python-dotenv==1.0.1
pyusb==1.2.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
safetensors==0.4.2
smbus==1.1.post2
soupsieve==2.5
spidev==3.6
sysv-ipc==1.1.0
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.39.3
twilio==9.0.2
typing_extensions==4.10.0
urllib3==2.2.1
validators==0.26.0
yarl==1.9.4
